Added support for badges in TabBar icons.
Added support for TabBar icon badges using the attribute badge in the
tab directive.

Ex: <tab icon-on="icon ion-ios7-person" icon-off="icon
ion-ios7-person-outline" href="#/tab/myprofile" badge="3">

I’ve used the colors of the TabBar itself but inverted for the badge. I
don’t know if you’ll want to merge this straight on but I think it can
help as a start point for this feature. I’ve tried to do it as
integrated as possible.