Clearify Description
There should be a more clear explanation of the percentages for the item width. It might be good to also have a visual of what comes from the grid usage code. When I try to click on the js or html links on the phone simulation the page either continues to wait or crashes (chrome wait or kill).