fix(utils/dom): do not add rAF polyfill to window
A pull request (#1430) was merged that added a requestAnimationFrame polyfill to
the window on platforms that do not have rAF.  This was overlooked.
This now has created conflicts with Angular 1.2.17 on Android browsers
that do not have requestAnimationFrame.

The problem was that a polyfill for requestAnimationFrame was put onto the
window for Android <4.3. AngularJS would then check if
window.requestAnimationFrame existed.  Angular's check passed, and then
it would try to use cancelAnimationFrame which was undefined.

Now, nothing on the window is changed.