[Prototype] Disable certain 'Back' button anims
Back button show/hide should not be animated in some situations when
switching to a different view (e.g. with a different history stack).

Currently the Title changes and Back Button appears/disappears at
different speeds, which is inconsistent. This change immediately
shows/hides the back button when an instant transition is required.