updated documentation to match code
Coming from the documentation `$ionicGesture.off` had to be called like `$ionicGesture.off('swipe', mySwipe, myElement)`. Instead, it has to be called with the gesture returned from `$ionicGesture.on` and the type+callback to remove.

I added the `@returns {ionic.Gesture} ...` to `.on` to make clear that this returns the gesture you may want to remove later on. I also made the documentation for `.off` match the function signature.