fix(collectionRepeat): rerender after resize/enter
Previously on every `beforeEnter`, collection-repeats within a
ion-nav-view would rerender. However, this is only necessary when the
window resizes. The rerender already works when the collection-repeat
is within the active view because the scroll view has accurate
dimensions. But when the collection-repeat is within a cached view it
does not have dimensions, causing the rerender to incorrectly place its
items. This update will only rerender the collection-repeat if there
was a window resize, and the scroll view which the collection-repeat
was in, did not have dimensions at the time of the resize. If/when the
view becomes the active view again, the collection-repeat will rerender
on `afterEnter` when the scroll view has accurate dimensions.