fix(cache): do not cache view after history change
If a view is set to not be cached, and it is the first view that was
loaded, and the next navigation was to a different history with its own
ion-nav-views (tabs), then remove the first view. Also allow for the
cache state property to be a string ‘false’ and boolean false.