Disable a single step of history recording for the next view displayed by specifying nextViewOptions.disableHistory.
This is useful when a root view takes you forward, but subsequent steps forward from the child should return to the root.

An example of this is:

Root is a list of available messages.  Tapping one takes you to a message display, which also contains a list of related messages.  If you record history and keep stepping through children you end up with a crazy history stack which is not what the user wants.  Pressing Back should return them to the root of the messages.

Adding this to nextViewOptions as disableHistory especially helps when using URL route switches to facilitate dynamic loading of states from the file system.