iOS menu-side-shadow fix
When testing an app with a side menu on iOS, the side-shadow does not appear. It seems iOS has an issue with multiple values for box-shadow, so I've opted instead create a single shadow that extends equally on both sides.

It should look identical to the existing shadow on most devices, but fix the issue with iOS displaying no shadow.

For reference, the device this was tested on was an iPhone 6 Plus.