Fix preLink function using compile attributes
The preLink function is using the compile attr not the $attr variable for the delegate-handle and locking attributes. This causes problems when using ion-content in an ng-repeat. The delegate-handle is properly set using the preLink attributes in other directives like ion-scroll. This does not address the issue of interpolation for delegate-handle but allows an additional directive to do so.