Update navBarDelegate to support bar appearance change
Fixes #2732

When I was developing a app I needed to change the navbar color depending on the screen. I tried using the rootScope workaround just like on issue #2732 but it didn't worked out because the navbar does not sync with it's inner headerBar. Besides that I feeled that this was not the best way to do this. I already develop native iOS apps and the navbar has methods to change it's appearance, that why I thought that the best solution is to add the ```appearance``` method to change it.