- fixed ion re-ordering support for collection-repeat Problem:   Note, <ion-reorder-button> works with ng-repeat. However, it is not   compatible with collection-repeat. The ion-re-ordering feature is based on the following assumptions:   1. style offset top attributes have been set DOM objects in the list items   2. that the aggregate of client heights can be summed
Complication:
  Fixing the re-ordering defect is highly problematic because:
  1. The ion-reordering feature is subject to filters which when applied
to the collection can affect one’s reordering callback.
  2. All items may not display with the same fixed client height


Solution:
  1. Establish the contract that the fromIndex & toIndex values sent to
the reordering callback must be values that are applicable to the
context of the current [filtered|orderedBy] view ONLY. In this manner it
is up to the developer to transform the from/to index values prior to
reordering the original data source
  a. Create separate functions for handling both hg-repeat & collection
repeat:
    - ng-repeat: uses: _getNgRepeatReorderIndex
    - collection-repeat: uses: _getColRepeatReorderIndex


  2. The ion-reordering for collection-repeat assumes a constant client
height for all items. However, if this is not the case then this renders
this feature useless and developers are forced to use ng-repeat. In such
instances an alternative approach would be to invoke a re-order callback
with a repeatReorder object. This object is initialised via:
  ReorderDrag.prototype._setRepeatReorderFromToItem = function(src, dst)
{
  	var scopeSrc = angular.element(src).scope();
  	var scopeDst = angular.element(dst).scope();
  	var scopeToUse = angular.isDefined(scopeDst.$repeatReorder) ?
scopeDst : scopeSrc;
  	scopeSrc.$repeatReorder.fromItem =
scopeSrc[scopeSrc.$repeatReorder.keyExpression];
  	scopeSrc.$repeatReorder.toItem =
scopeToUse[scopeSrc.$repeatReorder.keyExpression];
  }


Resulting context:
  1) using index values on ordering:
Eg.
<ion-list show-reorder="true"> 
  <ion-item collection-repeat="item in contacts | filter:data.search">
  …
  <ion-reorder-button class="opt-btn button-small icon ion-ios-drag"
on-reorder="reorderItem(item,$fromIndex,$toIndex)"></ion-reorder-button>

  2) using repeatReorder object on ordering:
<ion-list show-reorder="true"> 
  <ion-item collection-repeat="item in contacts | filter:data.search">
  …
  <ion-reorder-button class="opt-btn button-small icon ion-ios-drag"
on-reorder="reorderItem($repeatReorder)"></ion-reorder-button>


Known issues:
- Sometimes the drag item is lost when doing a move that spans multiple
scrolled pages. Hence, collection-repeat forces a re-load and the drag
action is lost. 
- When moving an item downwards the CSS z-index of the dragged item is
not being honoured and appears hidden behind other items in the list. My
work-around for this is to have items that have a transparent
background.