Do not prevent touchEvent if keyboard is not open
As the JSdoc of keyboardPreventDefault() states, the purpose is to prevent native scrolling while "the keyboard is open".

The function keyboardHide(), responsible for removing the listener on "touchMove" (which use keyboardPreventDefault() as a callback) may not be called when the keyboard is automatically closed (i.e: after submitting a form thanks to the "Go" button of the native keyboard of Android devices). As a result, the listener on "touchMove" may not be removed and keyboardPreventDefault may still be called on a different view (which does need the touchMove event... But this last will be Prevented).

Checking that the keyboard is open before preventing the event solves this issue.

NB: Calling close() method from the keyboard plugin does hide the keyboard, but does not execute keyboardHide().