Adding variable to store continuous intension
Adding a variable to store the original continuous intension, in the situation where a slidebox is used and  the slides are generated from ng-repeat liked to async call. The setup function is run and as there are no slides continuous is set to false. Later the async call and ng-repeat resolve and the developer calls update, which calls setup. As continuous is now set to false, it remains that way even when the developers original intension was to have it be continuous.

This pull request stores the original value if setup is going to change it and then if a later setup has more than two slides it is set back to continuous.