perf(cards): remove translateZ from ion-card
Yes, oddly enough this makes a performance difference. The translateZ
was used as a trick to force ion-card’s rounded corners to take effect
on inner elements. However, on a page that has many ion-cards, it takes
relatively a lot of power on a low-end device to move them all during a
transition. It’s best to not put them on their own layer and let the
browser just move one thing.