Currently the following methods are used to determine if a tab is active:
Assuming our route is setup like so:

  $stateProvider
    .state('contact', {
      url: "/contact",
      views: {
        'contact-tab': {
          templateUrl: "templates/contact.html"
        }
      }
    })

1. Does the href attribute of the <ion-tab> directive match the $location.path e.g.

	<ion-tab href="#/contact">
	=
	#/contact

2. Does the nav name of the nested <ion-nav-view> within <ion-tab> match the current state nav view name e.g.

	<ion-tab>
		<ion-nav-view name="contact-tab" ...
	</ion-tab>

   'contact-tab' = 'contact-tab'

3. Does the ui-sref attribute of the <ion-tab> directive match the current state e.g

	<ion-tab ui-sref="contact">
		<ion-nav-view...
	</ion-tab>

   'contact' = current state 'contact'

But often it is desireable to have a nested state within a tab e.g. to take advantage of resolved data inheritance for nested states.
In this case we might want to setup our rates like this:

    .state('contact', {
      url: "/contact",
      abstract: true,
      views: {
        'contact-tab': {
          template: "<ion-nav-view name='contact'></ion-nav-view>"
        }
      }
    })
    .state('contact.pageOne', {
      url: "/one",
      views: {
        'contact': {
          templateUrl: "templates/contact-page-one.html"
        }
      }
    })
    .state('contact.pageTwo', {
      url: "/two",
      views: {
        'contact': {
          templateUrl: "templates/contact-page-two.html"
        }
      }
    })

...and our tab markup...

      <ion-tab title="Contact" icon="ion-ios-world" ui-sref="contact.pageOne">
        <ion-nav-view name="contact-tab"></ion-nav-view>
      </ion-tab>

And that works fine EXCEPT when we navigate to contact.pageTwo from a view within another tab. The
view lands up in the tab's <ion-nav-view> but the tab itself doesn't activate because it's
testing 'contact.pageTwo' against 'contact.pageOne'. See http://codepen.io/lufin/pen/YqNjoy

Of course you can solve this by not nesting the sub pages:

    .state('contact', {
      url: "/contact",
      abstract: true,
      views: {
        'contact-tab': {
          template: "<ion-nav-view name='contact'></ion-nav-view>"
        }
      }
    })
    .state('contactPageOne', {
      url: "/one",
      views: {
        'contact-tab': {
          templateUrl: "templates/contact-page-one.html"
        }
      }
    })
    .state('contactPageTwo', {
      url: "/two",
      views: {
        'contact-tab': {
          templateUrl: "templates/contact-page-two.html"
        }
      }
    })

but then the parent state 'contact' becomes useless and we can't add any resolved data that might be useful for both contact pages e.g.

    .state('contact', {
      url: "/contact",
      abstract: true,
      views: {
        'contact-tab': {
          template: "<ion-nav-view name='contact'></ion-nav-view>"
        }
      }
      resolve: {
        somethingUseful: function(aService) {
          ....
        }
      }
    })
    .state('contactPageOne', {
      url: "/one",
      views: {
        'contact-tab': {
          templateUrl: "templates/contact-page-one.html"
          controller: function(somethingUseful) {
            ....
          }
        }
      }
    })
    .state('contactPageTwo', {
      url: "/two",
      views: {
        'contact-tab': {
          templateUrl: "templates/contact-page-two.html"
          controller: function(somethingUseful) {
            ....
          }
        }
      }
    })

In this commit I've made a stab at one solution - a new attribute for <ion-tab> called 'view-parents' that allows
us to determine if the tab is active based on the presence of a parent within the current state e.g.

<ion-tab ui-sref="contact.pageOne" view-parents="contact">

If the current state is 'contact.pageOne' then we can find 'contact' and identify the state as a child within the tab and
set the tab to active.

Note: we can't just split the ui-sref and take the first element and use that to test because you often want to load
multiple state trees (with different roots) into the tab. For this reason multiple view parents can be defined in the attribute e.g.

<ion-tab ui-sref="contact.pageOne" view-parents="contact differentContactParent">