Update modal.js
The usage example shows the HTML template wrapped into a script tag. It seems since version 1.7.7 this is not valid any more. If the modal HTML template is wrapped into a script tag, only the backdrop will be shown, but not the modal itself. I've spent few hours and found the correct answer here: http://stackoverflow.com/questions/32552896/ionic-modal-not-showing#answer-33292269