fix(toolbar): add the mode to the inverse function for a toolbar
Use the mode instead of the sass variable to determine which color the
buttons in a toolbar should be based on the background color. Added
error messages when anything other than a color is passed and more
comments.

BREAKING CHANGES: ios mode `$toolbar-ios-button-color` now has a
default value of `color-contrast($colors-ios, $toolbar-ios-background,
ios)` which will evaluate to the primary color for light background
toolbars and white for dark background toolbars.

fixes #6364