Update documentation for ion-slide
Reading the documentation and following along left me lost as to why the .lockNextSlides() was not working, this was because it required a boolean parameter that was not reflected in the documentation. This PR adds the JSDoc @param to show that the function requires a boolean parameter. Does not include a definition for the single parameters so this may cause issue in the theme of the docs template.