feat(Searchbar) - expose Focus and Blur events
Added code to expose the blur and focus events of the input field via EventEmitters (same as "ionInput" emitter). 

An example use case for this: using searchbar for the Google Places autocomplete together with a native Google Maps from @ionic-native. One has to disable / enable the map taps based on whether the searchbar is in focus  or not (otherwise all clicks on the Google Places suggestions dropdown get hijacked by the map). Also, there was a question about this functionality on StackOverflow: http://stackoverflow.com/questions/43756290/how-to-detect-onfocus-and-blur-event-of-ion-searchbar-in-ionic2