Register multiple callbacks to modals
This changes allows for registering multiple listeners for onDidDismiss and onWillDismiss. It is backwards compatible, does not require any changes to registering. Although, there may be a change in some behaviour: if you register multiple callbacks for onDidDismiss or onWillDismiss, it would have been overwritten in the past (only last register will execute) and now it will execute all registered listeners.

This is useful for situations where you would want to build a provider (for example 'modalmanager') that serves a modal to another provider, where both providers want to register to onDidDismiss or onWillDismiss.