src/components/tabs/tabs.ts
I added a new part about "Dynamic Tab Selection" because it wasn't clear previously. There are a lot of forum that talk about how call a specific tab from any page but the solutions are always strange or dirty. I hope that this adding can is helpfull for other developer