https://github.com/ionic-team/ionic/issues/11690 fix
The bug is in swiper-zoom.ts#L512. The event s.touchEvents.move is not defined. Furthermore you passing the the object {passive: false, zone:false} as config to plt.registerListener. But the subscription require a non passive listener to prevent the event from bubble up to the slides controller that handle the page slide event.