Fixes a bug that prevents keyboard from closing after clear button has been pressed
(click) and (mousedown) events are both fired when pressing clear(X) button, this causes a range of issues including not allowing the keyboard to close after clear has been pressed. This is due to the fact that _shouldBlur is left in the false state after clear rather then being reset to true. Issue is as follows.
1.Clear is pressed
2.clearInput from (mousdown) fires _shouldBlur = false
3.inputBlurred fires from deselecting input _shouldBlur = true
4.clearInput from (click) firest _shouldBlur = false.