Fix(scrollbar) prevent keyboard from closing
(click) and (mousedown) events are both fired when pressing clear(X) button, this causes a range of issues including not allowing the keyboard to close after clear has been pressed. This is due to the fact that _shouldBlur is left in the false state after clear rather then being reset to true. Issue is as follows.

Short description of what this resolves:
This issue
1.Clear is pressed
2.clearInput from (mousdown) fires _shouldBlur = false
3.inputBlurred fires from deselecting input _shouldBlur = true
4.clearInput from (click) fires _shouldBlur = false
5.Next blur will now refocus input do to shouldBlur = false