fix(content): align getScrollElement signatures
Fixes #15141.

This commit aligns the different signatures for IonContent.getScrollElement.

The return type in the interface generated by Stencil was HTMLElement. But
the Angular proxy generated by Stencil was returning a promise for it after
awaiting for componentReady(). All the uses of this method were also
awaiting for componentReady().

So this commit now implements the method in the web component to return
a (immediate) promise for the HTMLElement. The implementations of the using
components (infinite-scroll, refresher, reorder-group and virtual-scroll)
are also changed accordingly.

Now, you need to do without a framework:
```
await contentEl.componentReady();
let scrollEl = await contentEl.getScrollElement();
```

And with Ionic/Angular:
```
let scrollEl = await content.getScrollElement();
```