fix(textarea): prevent ui thrashing for textarea with autogrow
By wrapping the textarea in a div and always updating this div to the height of the scrollheight of the textarea, we prevent the UI thrashing that happens when the textarea's height is reset with 'auto'.  Previously, reseting the height to auto and immediately setting the height to the scrollheight caused the textarea to shrink to 0px tall before being reset to the scrollhgith. Wrapping this in a div prevents this sudden change in height of the component and prevents the scroll to top effect.

closes #19193