fix(reorder-group): add 'context' property
The context property allows autoScroll in reorder-group, even if it was created inside a small div (a div that has a limited height, maybe with a styled scrollbar - run "reorder-group context test" to see a use case).
Now, you can just add: context="{divID}", which does all the work for you!
You are welcome to change the property name to a better one :)