Remove window reference from defineCustomElements
The parameter `window` is no longer required as per stencil documentation. In fact, the first parameter to this function is ignored in the generated bundle.

Removing this reference will allow ionic to be imported in non-browser environments and makes it easier to integrate ionic in frameworks such as Next.js. This contributes a partial fix towards #19975