Allow null in virtual scroll [items]
Using the async pipe is considered best practice for angular projects.

When setting `angularCompilerOptions.strictTemplates=true` in `tsconfig.json` the build fails because there is a type mismatch between async & [items]. This is because the return type of async can be `any[] | null` whereas [items] only accepts `any[] | undefined`. Adding null to the allowed types fixes the issue and allows async to be used with virtual scroll.