[Fix] Use last item in the Intersection API callback
There is a rare race condition you may experience on devices with slower CPU where the Intersection Observer API supplies multiple items to the callback, instead of only a single element which the previous code assumed it would get.  This changes takes the last item (which will be the most current information).