fix(angular): config.get not a function in karma (#22853)
When karma runs each test case the window object persists. Ionic does
some initial setup in angular first before handing over to core.

On each test case in karma angular bootstraps itself again, during this
bootstrap we set Ionic.config. On the first test case this is fine as Ionic
core is also bootstrapped but on the second test case this overwrites
legitimate config which causes parts of Ionic to fail as the config is
no longer in the expected format.

This fix works by only setting Ionic.config if it does not already
exist. If it does exist then we do not replace it with an object thats
not compatible with core.

There may be better ways to solve this issue but I didn't want to change
too much of ionic/angular so this felt like the safest change