feat(datetime): add display prop & numpad UI
We started to get increasingly negative
feedback for the timepicker we used in our app,
so I opted to make the UX better on Desktop.

The Display prop controls, wether you want the datetime
component to show picker wheels, or only text inputs.

Feel free to check out the 'display' test and try out the UI.