refactor(router): introduce ResolvedRouteChain
Prior to this PR ResolvedRoute was used for 2 different things:
1. To describe the flattened router tree (readRoutesNodes),
2. To describe a route match (findChainForIDs, findChainForSegments).

There is a fundamental difference between the 2 because in case #1 the parameters are not resolved,
as there is no url to extract the parameters from at that point. In case #2 we could resolve the
parameters from either the IDs or the path.

Splitting the concepts makes the mental model clearer.

One other difference is that chains used to contain copies of the route nodes instead of references to those nodes.
See the changes in flattenNode.

Switching to references has several benefits:
- it consumes less memory and less CPU has we don't have to copy any more,
- it makes it trivial to compare routes but checking node equality which will allow further improvement.

Note:
This PR still has the concept of (unresolved) RouteChain to represent the flattened route tree.
However there is no strict need to flatten the tree. findChainForIDs and findChainForSegments will be updated
in a follow up CL to work with the tree instead of a flattened tree. This will save further memory and CPU.