fix issue with sometimes-null refs on datetime component
This fixes an issue with the Datetime component where in certain
circumstances a `ref` on an element rendered by the datetime component
would end up erroneously set to `null`, necessitating a yucky
`document.querySelector` based workaround.

The issue arises because of how Stencil's virtual DOM implementation
handles reconciling the children of old and new nodes, and in particular
how it deals with situations in which the order of children changes but
some of the nodes are conserved.

The way this was affecting the datetime component was that in certain
circumstances a method which returns a `<div>` with a `ref` would change
position in the children of another node. In particular, this can happen
if the `presentation` prop changes, causing the return value of the
`renderDatetime` function to change in turn. The salient changes to
the return value of _that_ function which would change look like
going from:

```ts
[
  this.renderCalendar(mode),
  this.renderTime()
]
```

to

```ts
[
  this.renderTime(),
  this.renderCalendar(mode)
]
```

`renderCalendar` in turn calls `renderCalendarBody` which returns the
`<div>` with the offending `ref` on it.

Anyhow, in reconciling these changed children when re-rendering the
component things can go wrong. Stencil's algorithm in this case does a
"best effort" at identifying nodes which should be kept between
re-renders, but it does not do an exhaustive check _unless nodes have a
key attr_. So, more or less, what was happening was that in cases where
the `<div>` with the `ref` on it was going to be included in the next
rendered output from the component Stencil would nonetheless remove the
'old' node and create a new identical one for the next render. Then
there is also a race condition where the new node would be created
before the old one is removed, causing the `ref` to be briefly set to
the new value (when the new VNode is created) and then reset to `null`
(when the old one was destroyed).

The fix is to set the `key` attribute on the same `<div>` as the `ref`.
This allows Stencil to do an exhaustive check when reconciling old and
new children in order to not needlessly throw away the old VDom node,
and gets rid of the behavior where the `ref` would end up set to `null`
erroneously. It's possible (but I think unlikely) that this change will
result in slightly less GC pressure, but I did no testing of that.

closes ionic-team/stencil#3253
FW-901