chore(ci): create stencil-eval job
this commit generates a baseline configuration for running ionic's ci
workflow with the latest and greatest version of stencil. to initialize
the workflow, `stencil-eval.yml` is a copy/paste of `nightly.yml`. this
can be verified by running `diff nighly.yml stencil-eval.yml` from
within the `.github/workflows` directory from this commit