chore(ci): create build-core-stencil-eval action
this commit creates `.github/workflows/actions/build-core-stencil-eval`
with the intent on using this action for building ionic core with an
eval version of stencil. in order to make the review easier, this commit
simply copy/pastes the existing `build-core` action into a new
directory (as of this commit, the two files are identical and can be
confirmed with `diff`).

`stencil-eval.yml` is updated to run the new build action instead of
`build-core`